Move email routes
