+++ index.js
@@ -5,12 +5,15 @@ const path = require('path');
 const { initDatabase } = require('./lib/database');
 const { createIndexes } = require('./lib/models/pastes');
 const pastes = require('./lib/routes/pastes');
+const email = require('./lib/routes/email');
 
 const app = express();
 
 // Parse application/json for all request
 app.use(express.json());
+
 app.use('/api/pastes', pastes);
+app.use('/api/email', email);
 
 if (process.env.NODE_ENV === 'production') {
   // Serve any static files
